fix(admin): Fix issues related to refactor
